core changes

+++ contracts/ISSVRegistry.sol
@@ -46,7 +46,7 @@ interface ISSVRegistry {
      * @param publicKey Operator's public key. Will be used to encrypt secret shares of validators keys.
      * @param fee The fee which the operator charges for each block.
      */
-    function registerOperator(string calldata name, address ownerAddress, bytes calldata publicKey, uint256 fee) external returns (uint32);
+    function registerOperator(string calldata name, address ownerAddress, bytes calldata publicKey, uint64 fee) external returns (uint32);
 
     /**
      * @dev removes an operator.
@@ -61,7 +61,7 @@ interface ISSVRegistry {
      */
     function updateOperatorFee(
         uint32 operatorId,
-        uint256 fee
+        uint64 fee
     ) external;
 
     /**
@@ -162,7 +162,7 @@ interface ISSVRegistry {
      */
     function getOperatorFee(uint32 operatorId)
         external view
-        returns (uint256);
+        returns (uint64);
 
     /**
      * @dev Gets active validator count.
